// src/app/modules/financeiro/contasreceber/enums/financeiro-contasreceber-enums.model.ts
/**********************************************************************************************
Code generated with MKL Plug-in version: 7.0.0
Code generated at time stamp: 2019-07-15T07:31:32.718
Copyright: Kerubin - logokoch@gmail.com

WARNING: DO NOT CHANGE THIS CODE BECAUSE THE CHANGES WILL BE LOST IN THE NEXT CODE GENERATION.
***********************************************************************************************/

import { SelectItem } from 'primeng/api';

export interface FormaPagamento extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoReceitaDespesa extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoContaBancaria extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoPlanoContaFinanceiro extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoPessoa extends SelectItem {
	
  label?: string;
  value: any;
  
}